We don't need to rearrange the SPI buses.
I was wrong, partially because I didn't realize that my labelling
of SPI pins was ambigious. The signal that was sitting on a PWM pin
_isn't_ part of the SPI bus, its just a GPIO signal that I can
move to another pin. This frees up the pins I need for PWM audio
without requiring a wholesale rearrangement of the PCB design.

So let's do that instead. I've called it board version 0.5...